kernel/Makefile: Remove -Wno-error=unused-label
Previously `FAILABLE_FUNCTION` functions always had prologue and epilogue
macros to keep failable functions consistent, and defined a label needed
for catching errors. However, that also meant if a function never returned
any errors, but had to be defined as `FAILABLE_FUNCTION` because of
callback's type, that label ended up being unused, since there was no error
being thrown anywhere in the function.

But all of this changed with last commit, as `FAILABLE_FUNCTION` is gone.
So this compiler flag became unnecessary.