shell: Remove special shell sections
The original idea behind "shell sections" was to be able to unload in the future, when we finally get an userspace shell.
However, this created serveral problems:
- This required correctly marking EVERY symbol inside kernel/shell directory, otherwise it would leave certain parts of shell after unloaded.
- Or worse, some stuff used by shell may never go into that section. Even if it did, it would be very compiler-dependent.
- And the marking had to be done *correctly*. If you accidentally put SHELL_FUNC instead of SHELL_DATA, you'll get cryptic linker failure.

I finally determined this is not worth the effort anyway, because the OS finally gets userspace and boots into it, even if we can unload the shell code, it would still get loaded during boot anyway.
More effective solution would be something like loading shell as kernel module or something similar, or getting rid of kernel-level shell entirely(when we have stable enough userspace).
Since we are still in very early days of kernel land, let's not think about that for now.